# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/payments/transactions": {
      "get": {
        "operationId": "AdminPaymentsController_getAllTransactions",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPayments"
        ]
      }
    },
...
    "/api/v1/admin/payments/payouts": {
      "get": {
        "operationId": "AdminPaymentsController_getAllPayouts",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPayments"
        ]
      }
    },
...
    "/api/v1/admin/payments/payouts/{payoutId}": {
      "get": {
        "operationId": "AdminPaymentsController_getPayoutDetails",
        "parameters": [
          {
            "name": "payoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPayments"
        ]
      }
    },
    "/api/v1/admin/payments/payouts/{payoutId}/approve": {
      "post": {
        "operationId": "AdminPaymentsController_approvePayout",
        "parameters": [
          {
            "name": "payoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessPayoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminPayments"
        ]
      }
    },
    "/api/v1/admin/payments/payouts/{payoutId}/process": {
      "post": {
        "operationId": "AdminPaymentsController_processPayout",
        "parameters": [
          {
            "name": "payoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessPayoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminPayments"
        ]
      }
    },
    "/api/v1/admin/payments/payouts/{payoutId}/complete": {
      "post": {
        "operationId": "AdminPaymentsController_completePayout",
        "parameters": [
          {
            "name": "payoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletePayoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminPayments"
        ]
      }
    },
    "/api/v1/admin/payments/payouts/{payoutId}/reject": {
      "post": {
        "operationId": "AdminPaymentsController_rejectPayout",
        "parameters": [
          {
            "name": "payoutId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectPayoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminPayments"
        ]
      }
    },
    "/api/v1/admin/payments/wallet-refunds": {
      "get": {
        "operationId": "AdminPaymentsController_listWalletRefundRequests",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPayments"
        ]
      }
    },
    "/api/v1/admin/payments/wallet-refunds/{refundId}": {
      "get": {
        "operationId": "AdminPaymentsController_getWalletRefundDetails",
        "parameters": [
          {
            "name": "refundId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPayments"
        ]
      }
    },
    "/api/v1/admin/payments/wallet-refunds/{refundId}/process": {
      "post": {
        "operationId": "AdminPaymentsController_processWalletRefund",
        "parameters": [
          {
            "name": "refundId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessWalletRefundDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminPayments"
        ]
      }
    },
    "/api/v1/admin/payments/gift-cards": {
      "get": {
        "operationId": "AdminPaymentsController_listGiftCards",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPayments"
        ]
      },
      "post": {
        "operationId": "AdminPaymentsController_createGiftCard",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGiftCardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminPayments"
        ]
      }
    },
    "/api/v1/admin/payments/gift-cards/{code}": {
      "get": {
        "operationId": "AdminPaymentsController_getGiftCard",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPayments"
        ]
      }
    },
    "/api/v1/admin/payments/gift-cards/{code}/status": {
      "patch": {
        "operationId": "AdminPaymentsController_updateGiftCardStatus",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGiftCardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminPayments"
        ]
      }
    },
    "/api/v1/admin/payments/transactions/{transactionId}/refund-razorpay": {
      "post": {
        "operationId": "AdminPaymentsController_refundRazorpayTransaction",
        "parameters": [
          {
            "name": "transactionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AdminPayments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Project API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ProcessPayoutDto": {
        "type": "object",
        "properties": {}
      },
      "CompletePayoutDto": {
        "type": "object",
        "properties": {}
      },
      "RejectPayoutDto": {
        "type": "object",
        "properties": {}
      },
      "ProcessWalletRefundDto": {
        "type": "object",
        "properties": {}
      },
      "CreateGiftCardDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateGiftCardDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}